requirements.txt
Flask==1.0.2
Werkzeug<0.13.0,>=0.12.0
google-cloud-datastore==1.6.0
google-cloud-tasks==0.3.0
web3==4.0.0
python-twitter